// src/interfaces/IAction.ts
// ----------------------------------------------------------------------------
// ----------------------------------------------------------------------------
export interface IAction {
  /**
   * Name of Property
   */
  methodName: string;
  /**
   * Name of Container class
   */
  modelName: string;
}